move new classes out of the jtwitter package
